requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
dj-email-url==1.0.6
Django==4.2
django-autoslug==1.9.9
django-cache-url==3.4.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
environ==1.0
environs==9.5.0
marshmallow==3.19.0
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
